+++ mi/ext/commands/__init__.py
@@ -1,5 +1,4 @@
-from .bot import *
 from .cog import *
-from .context import *
 from .core import *
+from .bot import *
 from .view import *